deepmind wrap
